services/openmind_service/requirements.txt
# OpenMind AI Server - Dependências
# Servidor de IA para análise de imagens de produtos

# Web Framework
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
python-multipart>=0.0.9

# Processamento de Imagens
Pillow>=11.0.0

# HTTP Requests
requests>=2.32.0
httpx>=0.27.0

# Validação e Schemas
pydantic>=2.9.0
pydantic-settings>=2.5.0

# Utilitários
python-dotenv>=1.0.0

# Logging
loguru>=0.7.0

# Nota: OpenAI será um serviço separado em container Docker dedicado

